requirements.txt
streamlit>=1.31
tensorflow-cpu
opencv-python-headless
numpy
pillow
supabase
python-dotenv
gdown
pytz